minor fix to register_contents
To avoid the contents of the expression register to be called as a
variable.

E.g: The content of my expression register ("=) was `pfile`. Before
adding the 1 to getreg() the registers.nvim mapping was trying to call
it as a (non-existing) variable instead of show the string "pfile".

See `:help getreg()`