fix: multiline registers of whitespace are not hidden or trimmed
If the register is multiple lines long and only contains white space,
it is not deemed to be "hidden" if the character used to join multiple
lines into one is not a white space character. Similarly it is also not
trimmed. This is because the pattern `^%s*(.-)%s*` is used to
determine if the register is just whitespace, but this will not work if
`"\n"` is replaced with a non whitespace character.